add `generate` CLI command
